Implementação de logging em eventos críticos: criação, atualização e exclusão de usuários

+++ src/uploads/upload.controller.ts
@@ -16,10 +16,16 @@ import { diskStorage } from 'multer';
 import { extname } from 'path';
 import { JwtAuthGuard } from 'src/auth/strategies/jwt-auth.guard';
 import { UploadService } from './upload.service';
+import { LoggingService } from 'src/logging/logging.service';
+import { UtilityService } from 'src/utility/Utility.service';
 
 @Controller('upload')
 export class UploadController {
-  constructor(private readonly uploadService: UploadService) {}
+  constructor(
+    private readonly uploadService: UploadService,
+    private readonly loggingService: LoggingService,
+    private readonly utilityService: UtilityService,
+  ) {}
   @Put('logo')
   @UseGuards(JwtAuthGuard)
   @UseInterceptors(
@@ -38,35 +44,54 @@ export class UploadController {
     @Req() request: Request,
     @Res() response: Response,
   ): Promise<any> {
+    const user = request.user;
+    const { id }: any = user;
+
+    if (!user || !id || !id.data) {
+      this.loggingService.warning('Usuário não autenticado ou ID inválido');
+      return response
+        .status(HttpStatus.UNAUTHORIZED)
+        .json({ message: 'Usuário não autenticado.' });
+    }
+    const uuidBuffer = Buffer.from(id.data);
     try {
-      const { id }: any = request.user;
-      const uuidBuffer = Buffer.from(id.data);
       const oldPathLogo = request.body.old_path_logo;
       await this.uploadService.updateImageUser(uuidBuffer, file.path, null);
 
       if (oldPathLogo) {
         fs.unlink(oldPathLogo, (error) => {
           if (error) {
-            console.error(error);
+            this.loggingService.error(
+              `Falha ao deletar logo ${oldPathLogo} em ${new Date().toISOString()}: ${error.message}`,
+            );
           } else {
             console.log('file deleted');
           }
         });
       }
 
+      this.loggingService.info(
+        `Logo ${file.path} atualizada com sucesso pelo usuário de id ${this.utilityService.bufferToUuid(uuidBuffer)} em ${new Date().toISOString()}`,
+      );
+
       return response.status(200).json({ message: 'Upload feito com sucesso' });
     } catch (error) {
       if (file) {
-        fs.unlink(file.path, function (err) {
-          if (err) {
-            console.log(err);
+        fs.unlink(file.path, function (error) {
+          if (error) {
+            this.loggingService.error(
+              `Falha ao deletar logo ${file.path} em ${new Date().toISOString()}: ${error.message}`,
+            );
           } else {
             console.log('File deleted!');
           }
         });
       }
+      this.loggingService.error(
+        `Falha ao atualizar logo ${file.path} pelo usuário de id ${this.utilityService.bufferToUuid(uuidBuffer)} em ${new Date().toISOString()}: ${error.message}`,
+      );
       throw new HttpException(
-        'Um erro interno ocorreu... Tente novamente mais tarde.',
+        'Falha ao atualizar logo. Por favor, tente novamente mais tarde.',
         HttpStatus.INTERNAL_SERVER_ERROR,
       );
     }
@@ -89,19 +114,27 @@ export class UploadController {
     @Req() request: Request,
     @Res() response: Response,
   ): Promise<any> {
-    try {
-      const { id }: any = request.user;
-      const uuidBuffer = Buffer.from(id.data);
-      const oldPathCover = request.body.old_path_cover;
+    const user = request.user;
+    const { id }: any = user;
 
-      console.log('oldVover:? ', oldPathCover);
+    if (!user || !id || !id.data) {
+      this.loggingService.warning('Usuário não autenticado ou ID inválido');
+      return response
+        .status(HttpStatus.UNAUTHORIZED)
+        .json({ message: 'Usuário não autenticado.' });
+    }
+    const uuidBuffer = Buffer.from(id.data);
+    try {
+      const oldPathCover = request.body.old_path_cover;
 
       await this.uploadService.updateImageUser(uuidBuffer, null, file.path);
 
       if (oldPathCover) {
         fs.unlink(oldPathCover, (error) => {
           if (error) {
-            console.error(error);
+            this.loggingService.error(
+              `Falha ao deletar capa ${oldPathCover} em ${new Date().toISOString()}: ${error.message}`,
+            );
           } else {
             console.log('file deleted');
           }
@@ -113,14 +146,19 @@ export class UploadController {
       if (file) {
         fs.unlink(file.path, (error) => {
           if (error) {
-            console.error(error);
+            this.loggingService.error(
+              `Falha ao deletar capa ${file.path} em ${new Date().toISOString()}: ${error.message}`,
+            );
           } else {
             console.log('file deleted');
           }
         });
       }
+      this.loggingService.error(
+        `Falha ao atualizar capa ${file.path} pelo usuário de id ${this.utilityService.bufferToUuid(uuidBuffer)} em ${new Date().toISOString()}: ${error.message}`,
+      );
       throw new HttpException(
-        'Um erro interno ocorreu... Tente novamente mais tarde.',
+        'Falha ao atualizar capa. Por favor, tente novamente mais tarde.',
         HttpStatus.INTERNAL_SERVER_ERROR,
       );
     }
